+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/MediaPeriodHolder.java
@@ -65,8 +65,8 @@ import androidx.media3.exoplayer.upstream.Allocator;
    *
    * <p>Renderers that are needed must have been enabled with the {@link #sampleStreams} for this
    * {@link #mediaPeriod}. This means either {@link Renderer#enable(RendererConfiguration, Format[],
-   * SampleStream, long, boolean, boolean, long, long)} or {@link Renderer#replaceStream(Format[],
-   * SampleStream, long, long)} has been called.
+   * SampleStream, long, boolean, boolean, long, long, MediaPeriodId)} or {@link
+   * Renderer#replaceStream(Format[], SampleStream, long, long, MediaPeriodId)} has been called.
    *
    * <p>Renderers that are not needed must have been {@link Renderer#disable() disabled}.
    */